detect stale reviews and submit them periodically (#66)
we defined a number of metrix for treating a review as an eligible
for automatic submit. also we need to submit forcibly reviews which
were not submitted before from a background job.
there are two tests: for gating and for submit. but they have similar
structure. let's implement them using the template method pattern
and pass the test strategy as a parameter for the generic algorithm.

Co-authored-by: root <root@centos7.linuxvmimages.local>