enable keep alives for long ssh sessions
broken ssh connections are observed. a possible reason may be
disconnects by provider network. let's use the keep alive feature
of the ssh protocol and see if there will be any improvements.
need to patch SSH_OPTIONS value in jenkins as it provides an
explicit variable value and overrides the one in tf-devstack.

Change-Id: I6c4ae184a2c1558a6ac40392046b25b344ba9298